fix: change buyQuantity/sellQuantity from Long to Double
Original ReportEntity had this exact bug (quantity was Long but source
was double). This caused fractional quantities from stock splits,
demergers, and mutual funds to be truncated. e.g. 5.5 -> 5.

Updated:
- TradeOutcomeEntity buyQuantity/sellQuantity: Long -> Double
- TradeOutcomeContext buyQuantity/sellQuantity: Long -> Double
- PortfolioService: removed (long) cast in toTradeOutcomeContext
- TradeOutcomeMigration: parseLong -> parseDouble for sell_quantity
- AnalyticsIntegrationTest: 10L -> 10.0

+++ backend/src/main/java/com/thiru/investment_tracker/dto/context/TradeOutcomeContext.java
@@ -32,14 +32,14 @@ public class TradeOutcomeContext {
     // Buy side
     private double originalBuyPrice;
     private double caAdjustedBuyPrice;
-    private Long buyQuantity;
+    private Double buyQuantity;
     private LocalDate buyDate;
     private double buyBrokerCharges;
     private double buyMiscCharges;
 
     // Sell side
     private double sellPrice;
-    private Long sellQuantity;
+    private Double sellQuantity;
     private LocalDate sellDate;
     private double sellBrokerCharges;
     private double sellMiscCharges;
